remapped variables in model from Integers to Doubles

+++ app/src/main/java/org/wit/wishlistandroid/models/NodeModel.kt
@@ -9,9 +9,9 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class NodeModel(var id: Long = 0,
                      var title: String = "",
-                     var minTemp: Int = 0,
-                     var maxTemp: Int = 0,
-                     var minHumid: Int = 0,
-                     var maxHumid: Int = 0,
-                     var minMoisture: Int = 0,
-                     var maxMoisture: Int = 0) : Parcelable
+                     var minTemp: Double = 0.0,
+                     var maxTemp: Double = 0.0,
+                     var minHumid: Double = 0.0,
+                     var maxHumid: Double = 0.0,
+                     var minMoisture: Double = 0.0,
+                     var maxMoisture: Double = 0.0) : Parcelable
